sms_credit plugin: fix wrong password when creating user
sms_credit plugin permits to create a new user in Kalkun.

The code to create the user was not up to date and diverged from the one in
User_model.

Now sms_credit uses the code in User_model to avoid duplication of code

Same change was applied user deletion through sms_credit plugin