add contact: close underlying dialog when openning the new one
When adding contact in the phonebook, a 1st dialog appears asking whether to
create a signe contact or import from CSV. When chosing one, the dialog remained
in the background, and when addition was finished, it was still there what
is annoying. So close it when we open the new one.

+ increase height of CSV import dialog. It was a bit too short